pyproject.toml
[tool.poetry]
name = "lesson-28-hw-django"
version = "0.1.0"
description = ""
authors = ["VladmimirShtefan <shtefanvladimir14@gmail.com>"]
readme = "README.md"
packages = [{include = "lesson_28_hw_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
pillow = "^9.3.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
